# ⚡️ Speed up function `check_for_nltk_package` by 111%